concat decompression-reader.go in read.go and keep the original code mentions
